fix: throw error on getField on new create page
The entry data is empty when the user is on the create new page. This
would work if getField is used to fetch root elements. But, when we try
to fetch the nested elements it fails with no useful information to the
user. Hence, in order to have a consistent output, we are throwing an
error for all the getField() calls within the new entry page.

We determine whether the page is in create mode or edit mode by
checking if the entry is present. Because entry data is not passed to
SDK when it is in create page.

Please Note: Since, we allow only custom fields within the create new
entry page, this change only applys to the custom fields.

fixes: CS-27725